+++ lib/td_cache/acl_cache.ex
@@ -62,6 +62,24 @@ defmodule TdCache.AclCache do
     end
   end
 
+  def get_acl_role_users_multikey(resources, role) do
+    keys =
+      Enum.flat_map(resources, fn {resource_type, resource_ids} ->
+        Enum.map(resource_ids, fn resource_id ->
+          Keys.acl_role_users_key(resource_type, resource_id, role)
+        end)
+      end)
+
+    [
+      ["SUNIONSTORE", "union_list"] ++ keys,
+      ["SINTER", "union_list", UserCache.ids_key()]
+    ]
+    |> Redix.transaction_pipeline!()
+    |> case do
+      [_, user_ids] -> Redix.to_integer_list!(user_ids)
+    end
+  end
+
   def has_role?(resource_type, resource_id, role, user_id) do
     key = Keys.acl_role_users_key(resource_type, resource_id, role)