+++ test/td_cache/acl_cache_test.exs
@@ -22,6 +22,8 @@ defmodule TdCache.AclCacheTest do
       Redix.del!([
         "acl_roles:test_type:*",
         "acl_role_users:test_type:*",
+        "acl_role_users:domain:*",
+        "acl_role_users:structure:*",
         "acl_group_roles:test_type:*",
         "acl_role_groups:test_type:*",
         "permission:foo:roles",
@@ -82,6 +84,35 @@ defmodule TdCache.AclCacheTest do
 
     users_result = AclCache.get_acl_role_users(@resource_type, @resource_id, @role)
     assert user_id in users_result
+  end
+
+  test "get_acl_role_users_multikey for multiple domains and structures keys" do
+    role = "multikey_role"
+    domain_id = System.unique_integer([:positive])
+    structure_id = System.unique_integer([:positive])
+
+    domain_user_ids = System.unique_integer([:positive])
+    structure_user_ids = System.unique_integer([:positive])
+    CacheHelpers.put_user_ids([domain_user_ids, structure_user_ids])
+
+    invalid_domain_user_ids = System.unique_integer([:positive])
+    invalid_structure_user_ids = System.unique_integer([:positive])
+
+    AclCache.set_acl_role_users("domain", domain_id, role, [
+      domain_user_ids,
+      invalid_domain_user_ids
+    ])
+
+    AclCache.set_acl_role_users("structure", structure_id, role, [
+      structure_user_ids,
+      invalid_structure_user_ids
+    ])
+
+    assert [domain_user_ids, structure_user_ids] ==
+             AclCache.get_acl_role_users_multikey(
+               %{"domain" => [domain_id], "structure" => [structure_id]},
+               role
+             )
   end
 
   test "delete_acl_role_users deletes from cache" do